add the mobile support using walletConnect
